+++ app/controllers/tasks/field_occurrences/browse_controller.rb
@@ -1,0 +1,4 @@
+class Tasks::FieldOccurrences::BrowseController < ApplicationController
+  include TaskControllerConfiguration
+
+end