+++ CHANGELOG.md
@@ -9,6 +9,59 @@ This project <em>does not yet</em> adheres to [Semantic Versioning](https://semv
 
 ### Added
 
+- Append mode to multi update data attributes task [#4204]
+- <in progress Batch add/remove Protocols>
+- Protocol `is_machine_output` attribute to facilitate MachineObservation types
+- Sticky navbar on browse nomenclature and collecting event
+- Move button to citations and depiction slices in radial annotator
+- Freeze column checkbox to filter tasks [#4220]
+- Sort layout columns on filter tasks [#4219]
+
+### Changed
+
+- Add total to unmatched panel on Nomenclature match task
+
+### Fixed
+
+- Rendering of Identifier::Local::RecordNumber
+- Reset button is not working on Unify people
+
+[#4204]: https://github.com/SpeciesFileGroup/taxonworks/pull/4204
+[#4219]: https://github.com/SpeciesFileGroup/taxonworks/pull/4219
+[#4220]: https://github.com/SpeciesFileGroup/taxonworks/pull/4220
+
+## [0.48.0] - 2025-02-14
+
+### Added
+
+- Browse Field Occurrence task [#4200]
+- Help for Project Vocabulary task UI fields [#4192]
+
+### Changed
+
+- Sort asserted distributions alphabetically in Quick Forms
+- Update taxon nomenclatural date after source update
+- Updated Ruby Gems
+
+### Fixed
+
+- Failure to set `dwc_occurrences` for rebuild for background processing
+- New collecting event task: Identifier panel sets and displays an incorrect identifier number [#4208]
+- Comprehensive digitization sometimes failing to save identifiers [#4206]
+- Bad link when creating a synonym where the old name has children and then clicking the green edit button in Edit Taxon Name [#4196]
+- Taxon names not being displayed in relationships facet of Filter Nomenclature task. [#4193]
+
+[#4208]: https://github.com/SpeciesFileGroup/taxonworks/issues/4208
+[#4206]: https://github.com/SpeciesFileGroup/taxonworks/pull/4206
+[#4196]: https://github.com/SpeciesFileGroup/taxonworks/pull/4196
+[#4193]: https://github.com/SpeciesFileGroup/taxonworks/pull/4193
+[#4192]: https://github.com/SpeciesFileGroup/taxonworks/pull/4192
+[#4200]: https://github.com/SpeciesFileGroup/taxonworks/pull/4200
+
+## [0.47.0] - 2025-02-06
+
+### Added
+
 - Use Rails 7.2 and Ruby 3.3.6
 - New image task: Add field occurrence to depict some list [#4135]
 - Grab cursor to make sorting feature visible [#4153]
@@ -34,11 +87,11 @@ This project <em>does not yet</em> adheres to [Semantic Versioning](https://semv
 - Match identifiers defaults to match Identifier, not internal, `\n`, and caseless match [#4176]
 - Added Gift status notices to loans form [#4174]
 - Improved and clarified DwcOccurrence indexing concepts and application
-- Improved writing to cached* fields for TaxonName 
+- Improved writing to cached\* fields for TaxonName
 
 ### Fixed
 
-- Fixes to TaxonWorks CSL style 
+- Fixes to TaxonWorks CSL style
 - Bug in Image autocomplete
 - CatalogNumbers attached to Containers not appearing in CollectionObject tag [#4163]
 - OriginRelationship creation for Sequences [#4180]
@@ -51,7 +104,6 @@ This project <em>does not yet</em> adheres to [Semantic Versioning](https://semv
 - Images added before saving field occurrence are not saved [#4134]
 - Rendering Family group names from invalid names [#4187]
 - Verbatim latitude not displaying [#4178]
-
 
 [#4056]: https://github.com/SpeciesFileGroup/taxonworks/issues/4056
 [#4177]: https://github.com/SpeciesFileGroup/taxonworks/issues/4177
@@ -5072,7 +5124,9 @@ _Special thanks to Tom Klein for his amazing open-source contributions on this r
 - Loosing input page numbers when switching tabs on New Taxon Name task
 
 [#1532]: https://github.com/SpeciesFileGroup/taxonworks/issues/1532
-[unreleased]: https://github.com/SpeciesFileGroup/taxonworks/compare/v0.46.1..development
+[unreleased]: https://github.com/SpeciesFileGroup/taxonworks/compare/v0.48.0...development
+[0.48.0]: https://github.com/SpeciesFileGroup/taxonworks/compare/v0.47.0...v0.48.0
+[0.47.0]: https://github.com/SpeciesFileGroup/taxonworks/compare/v0.46.1...v0.47.0
 [0.46.1]: https://github.com/SpeciesFileGroup/taxonworks/compare/v0.46.0...v0.46.1
 [0.46.0]: https://github.com/SpeciesFileGroup/taxonworks/compare/v0.45.0...v0.46.0
 [0.45.0]: https://github.com/SpeciesFileGroup/taxonworks/compare/v0.44.3...v0.45.0